made querying the players in leagues easier from Sleeper & implemented multiple screens using a bottom bar

+++ app/src/main/java/com/amped94/ffbtracker/data/model/db/dao/UserDao.kt
@@ -10,7 +10,7 @@ interface UserDao {
     suspend fun getAll(): List<User>
 
     @Query("SELECT * FROM User WHERE userId = :id")
-    suspend fun getUser(id: Int): User?
+    suspend fun getUser(id: Long): User?
 
     @Query("SELECT * FROM User WHERE username = :username")
     suspend fun getUser(username: String): User?
